Added new carbon starvation model.
The new approach is similar to the method proposed by Camac et al. (2018), except that in
here we use carbon storage relative to target as the dependent variable. This choice is
preferred over growth rate because growth can be zero for deciduous trees even plants are
leafless, even though the plant may have plenty of storage.